requirements.txt
aiosqlite
bcrypt
chromadb
docling[ocrmac]
fastapi
lancedb
ollama
pydantic
pydantic-settings
python-dotenv
python-multipart
sqlalchemy[asyncio]
uvicorn[standard]